refactor: view - logic 분리
